backend/requirements.txt
fastapi
uvicorn
web3
solana
python-telegram-bot==20.3
httpx